Restore ensemble docstrings and freeze configs
